requirements.txt
dj-database-url==0.5.0
Django==2.2
django-crispy-forms==1.11.1
django-heroku==0.3.1
django-six==1.0.4
djangorestframework==3.12.2
gunicorn==20.0.4
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
xlwt==1.3.0
